feat: create slice interface
